SysLink: Samples - Various fixes to the Notify sample
	1. Calls the UsrUtilsDrv_setup() to set up
	   Memory_OS module.
	2. Use UInt32 for proc_id to match with kernel
	   side usage.
	3. Clean up the notify ping sample to output
	   fewer logs.
	4. Clean up the formatting of the overall sample.

+++ syslink/api/include/NotifyDrvDefs.h
@@ -123,7 +123,7 @@ typedef struct NotifyDrv_EventPacket_tag {
     /*!< List element header */
     UInt32             pid;
     /* Process identifier */
-    UInt16             procId;
+    UInt32             procId;
     /*!< Processor identifier */
     UInt32             eventNo;
     /*!< Event number used for the registration */
